Add files via upload
The project is written in Python. 1) Connect to a Microsoft SQL Server database. 2) Using SQL queries, we get data (average and maximum for given columns for a given period) 3) Using the Pandas library, save the data in a DataFrame, and then write the data to Excel 4) OK, I received the Excel file I needed with the necessary data